+++ original/CWP-Generate/CWPWideScaling.py
