+++ src/bpmncwpverify/core/typechecking.py
